requirements.txt
pandas
matplotlib
seaborn
scikit-learn
joblib
djangorestframework 